Merge branch 'branch-A-CodeQuality'
Change variable names

There are several variables that i have changed and style because of the comments that reviwers gave before and i ammended them accordingly.
Based on the comments, i further checked my other files and change them accordingly.

Also, the some of the methods and that were having complicated expressions were edited to be simpler for better understanding and readability.